backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
python-docx==1.2.0
docxtpl==0.20.2
django-cors-headers==4.9.0
gunicorn==25.1.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
